+++ AdjusterMain.py
@@ -24,7 +24,8 @@ def adjust(args):
     if not hasattr(args,"sprite"):
         args.sprite = None
 
-    apply_rom_settings(rom, args.heartbeep, args.heartcolor, args.quickswap, args.fastmenu, args.disablemusic, args.sprite, args.ow_palettes, args.uw_palettes)
+    apply_rom_settings(rom, args.heartbeep, args.heartcolor, args.quickswap, args.fastmenu, args.disablemusic,
+                       args.sprite, args.ow_palettes, args.uw_palettes, args.reduce_flashing)
 
     rom.write_to_file(output_path('%s.sfc' % outfilebase))