+++ Main.py
@@ -27,7 +27,7 @@ from Fill import sell_potions, sell_keys, balance_multiworld_progression, balanc
 from ItemList import generate_itempool, difficulties, fill_prizes, customize_shops
 from Utils import output_path, parse_player_names
 
-__version__ = '0.4.0.6-u'
+__version__ = '0.4.0.7-u'
 
 
 class EnemizerError(RuntimeError):
@@ -280,7 +280,9 @@ def main(args, seed=None, fish=None):
                 rom_names.append((player, team, list(rom.name)))
                 world.spoiler.hashes[(player, team)] = get_hash_string(rom.hash)
 
-                apply_rom_settings(rom, args.heartbeep[player], args.heartcolor[player], args.quickswap[player], args.fastmenu[player], args.disablemusic[player], args.sprite[player], args.ow_palettes[player], args.uw_palettes[player])
+                apply_rom_settings(rom, args.heartbeep[player], args.heartcolor[player], args.quickswap[player],
+                                   args.fastmenu[player], args.disablemusic[player], args.sprite[player],
+                                   args.ow_palettes[player], args.uw_palettes[player], args.reduce_flashing[player])
 
                 if args.jsonout:
                     jsonout[f'patch_t{team}_p{player}'] = rom.patches